I support inputs with fewer dimensions than the prototype
It's possible for my arguments (and the output) to have fewer dimensions than
required by the prototype, and still pass all the dimensionality checks,
assuming implied leading dimensions of length 1. For instance I could receive a
scalar where a ('n',) dimension is expected, or a ('n',) vector where an
('m','n') array is expected. I thus make a local (and padded) copy of the
strides and dims arrays, and use those where needed. Most of the time these will
just be copies of the input. The dimension counts and argument counts will be
relatively small, so this is only a tiny bit wasteful