c broadcasting: I handle output dimensions early
This is a big and mostly no-op patch. It does 99% of the work to allow new named
dimensions in outputs, to be defined by the passed-in output arrays. A big logic
difference: broadcasted slices in the outputs must match the shape implied by
the inputs EXACTLY. No implicit leading length-1 dimensions in the outputs. No
extra slices in the outputs only. No length-N slices in the outputs where the
inputs are all length 1.