even more precise logic for size-0 concatenation
I want this idiom to work:

  accum = np.array(())
  while ...:
     x     = ...
     accum = nps.glue(accum, x,  axis = -2)

np.array(()).shape is (0,), so the usual concatenation rules would make the
above fail. I want accum=x on the first pass in the common cases, and this patch
special-cases that logic.

This will potentially do the wrong thing if size.x == 0, but that's not a common
case.