broadcast_define(): better out_kwarg logic
If all of:

- out_kwarg is given in the broadcast definition
- output array is NOT given when called
- the output shape is available

Then we construct the output array before any slices are evaluated, and always
fill it in. This was already happening when broadcasting was actually active,
but not in a single-slice call.

Additionally, the single-slice path now makes sure to always return the output
array, if one is available. This is in place of whatever the inner function
actually returns, which may be bogus, if in-place output is enabled